streamlit/Infrastructure/requirements.txt
aws-cdk-lib>=2.99.0